requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# NLP & Transformers
transformers==4.30.2
torch==2.0.1
tokenizers==0.13.3
sentencepiece==0.1.99

# ONNX for optimization
onnx==1.14.0
onnxruntime==1.15.1

# Energy & Carbon Tracking
codecarbon==2.3.4

# API & Web Framework
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.3

# Dashboard
streamlit==1.25.0
plotly==5.15.0

# Database
sqlalchemy==2.0.19
alembic==1.11.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tqdm==4.65.0
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Data Processing
beautifulsoup4==4.12.2
html5lib==1.1
email-validator==2.0.0

# Model Calibration
netcal==1.3.5
